Fixed feature detection
When compiling an empty file, the compiler would fail when linking, because no main() existed (and it wasn't specified that it was a library). Because of this, the features were never found, because the compiler failed no matter what.
NOP (no target) time for the Makefile decreased by almost 89%, and it successfully detected both features.